Add env cluster selection and lua plugin (#1)
* Add a more modular structure for keymaps

* Auto format using gopls

* Fix coloring of selection

* First attempt at fetching output to nvim

* Add cluster id to list secondary item

This allows us to fetch it for piping out, without having to see it
visably

Add option to write selection to file

* Update flag to be nvim

* Add initial class with install and open

This is a lua package with a class to operate a floating terminal and
install the go package

TODO:
- Fix the install
- Add more docs
- prob more

* Fix install to run

This involves installing with make and go. Hence an initial test is done
to check if the dependencies exists, then the install is performed.